Add NameGen

+++ include/Keyboard.h
@@ -27,6 +27,6 @@ class Keyboard {
 	public:
 		static constexpr int MAX_LENGTH = 64;
 		static bool openForText(std::function<void(std::string)> f, const std::string &headerText = "", const std::string &subText = "", int maxStringLength = MAX_LENGTH, const std::string &initialText = "", u32 kbdDisableBitmask = 0);
-		static bool openForNumber(std::function<void(int)> f, const std::string &headerText = "", const std::string &subText = "", int maxStringLength = MAX_LENGTH, const std::string &initialText = "", const std::string &leftButton = "", const std::string &rightButton = "", u32 kbdDisableBitmask = 0);
+		static bool openForNumber(std::function<void(s64)> f, const std::string &headerText = "", const std::string &subText = "", int maxStringLength = MAX_LENGTH, const std::string &initialText = "", const std::string &leftButton = "", const std::string &rightButton = "", u32 kbdDisableBitmask = 0);
 		// Use SwkbdKeyDisableBitmask_* for kbdDisableBitmask.
 };
